requirements.txt
ipykernel
pandas
numpy
pydantic
pytest
matplotlib
plotly
seaborn
from-root
fastapi
uvicorn
python-dotenv
streamlit
mysql-connector-python
wordcloud
-e .
